REFACTOR - MOST OF IT
