+++ README.md
@@ -22,8 +22,8 @@ InstantNeo is a Python library that lets you create LLM-based agents quickly and
 ## Features
 
 - **Unified Provider Interface:** Switch seamlessly between models providers with consistent syntax. The library handles different API requirements behind the scenes.
-- **Quick and powerful Skill Management:** Define agent capabilities using simple Python decorators that transform functions into skills with metadata, descriptions, and parameter validation. Add, remove, and list skills dynamically as your agent's needs evolve.
-- **Flexible Execution Modes:** Control exactly how skills are executed with three modes: wait for results, fire skills in the background, or just extract arguments without execution for planning purposes.
+- **Quick and powerful Tool Management:** Define agent capabilities using simple Python decorators that transform functions into tools with metadata, descriptions, and parameter validation. Add, remove, and list tools dynamically as your agent's needs evolve.
+- **Flexible Execution Modes:** Control exactly how tools are executed with three modes: wait for results, fire tools in the background, or just extract arguments without execution for planning purposes.
 - **Text and Image Support:** Process both text and images through a single consistent API. Send images alongside prompts to vision-capable models and control the level of image analysis detail as needed.
 - **Customizable Agent Settings:** Modify agent behavior on-the-fly by overriding temperature, max tokens, role setup, and other parameters for specific interactions without recreating the entire agent.
 
@@ -55,10 +55,9 @@ print(neo.run("What's the Matrix?"))
 ### Teach Him Kung Fu
 
 ```python
-from instantneo.skills import skill
-from instantneo import InstantNeo
+from instantneo import InstantNeo, tool
 
-@skill(description="Execute a kung fu move", parameters={"move": "e.g., dragon punch", "intensity": "1-10"})
+@tool(description="Execute a kung fu move", parameters={"move": "e.g., dragon punch", "intensity": "1-10"})
 def kung_fu(move: str, intensity: int = 5) -> str:
     return f"Hit {move} at intensity {intensity}. I know kung fu!"
 
@@ -67,7 +66,7 @@ neo = InstantNeo(
     api_key="your-api-key",
     model="claude-3-7-sonnet-20250219",
     role_setup="You are Neo, martial arts downloaded.",
-    skills=["kung_fu"]
+    skills=[kung_fu]
 )
 
 print(neo.run("Three agents ahead. What move?"))
@@ -91,7 +90,7 @@ print(neo.run("I need an exit."))
 
 ## API Reference
 
-- **InstantNeo Class**: Set provider, api_key, model, role_setup, plus skills and tuning params (temperature, max_tokens, etc.).
+- **InstantNeo Class**: Set provider, api_key, model, role_setup, plus tools and tuning params (temperature, max_tokens, etc.).
 - **run() Method**: Feed it a prompt, pick an execution_mode (WAIT_RESPONSE, EXECUTION_ONLY, GET_ARGS), override settings as needed.
 
 Details in the docs.
@@ -102,7 +101,7 @@ InstantNeo leans on Minsky's "Society of Mind": small, specialized agents you co
 
 ## Documentation
 
-Full scoop at instantneo.readthedocs.io, including tutorials on multi-agent setups and advanced skills.
+Full scoop at /docs, including tutorials on multi-agent setups and advanced skills.
 
 ## Contributing
